Added Cart file and stored function there
