+++ src/App.tsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import AppHeader from './Title';
 import Menu from './Menu';
+import Cart from './Cart';
 import { Pokemon } from './Pokemon-Interface';
 
 function App() {
@@ -30,25 +31,6 @@ function App() {
         <Menu handleAdd={handleAdd}/>
         <Cart cart={cart} handleRemove={handleRemove}/>
       </>
-  )
-}
-
-interface CartProps {
-  cart: Pokemon[];
-  handleRemove: (poke: Pokemon)=>void;
-}
-
-function Cart({cart, handleRemove}: CartProps) {
-  return (
-    <>
-     <h2>Cart:</h2>
-     {cart.map(poke =>
-        <ul key={poke.id}>
-           <li>{poke.name}</li>
-           <button onClick={() => handleRemove(poke)}>remove</button>
-        </ul>
-      )}
-    </>
   )
 }
